Updated v6 with latest changes and added new isPartitioned flag to avoid error

+++ test/integration/query-interface.test.js
@@ -3,7 +3,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 const Support = require('./support');
-const DataTypes = require('../../lib/data-types');
+const DataTypes = require('sequelize/lib/data-types');
 const dialect = Support.getTestDialect();
 const Sequelize = Support.Sequelize;
 const current = Support.sequelize;
@@ -35,29 +35,38 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
 
   describe('showAllTables', () => {
     it('should not contain views', async function() {
-      async function cleanup() {
-        // NOTE: The syntax "DROP VIEW [IF EXISTS]"" is not part of the standard
-        // and might not be available on all RDBMSs. Therefore "DROP VIEW" is
-        // the compatible option, which can throw an error in case the VIEW does
-        // not exist. In case of error, it is ignored.
-        try {
-          await this.sequelize.query('DROP VIEW V_Fail');
-        } catch (error) {
-          // Ignore error.
+      async function cleanup(sequelize) {
+        if (dialect === 'db2') {
+          await sequelize.query('DROP VIEW V_Fail');
+        } else if (dialect === 'oracle') {
+          const plsql = [
+            'BEGIN',
+            'EXECUTE IMMEDIATE',
+            '\'DROP VIEW V_Fail\';',
+            'EXCEPTION WHEN OTHERS THEN',
+            '  IF SQLCODE != -942 THEN',
+            '    RAISE;',
+            '  END IF;',
+            'END;'
+          ].join(' ');
+          await sequelize.query(plsql);
+        } else {
+          await sequelize.query('DROP VIEW IF EXISTS V_Fail');
         }
       }
       await this.queryInterface.createTable('my_test_table', { name: DataTypes.STRING });
-      await cleanup();
-      await this.sequelize.query('CREATE VIEW V_Fail AS SELECT 1 Id');
+      await cleanup(this.sequelize);
+      const sql = dialect === 'db2' ? 'CREATE VIEW V_Fail AS SELECT 1 Id FROM SYSIBM.SYSDUMMY1' : `CREATE VIEW V_Fail AS SELECT 1 Id${  Support.addDualInSelect()}`;
+      await this.sequelize.query(sql);
       let tableNames = await this.queryInterface.showAllTables();
-      await cleanup();
+      await cleanup(this.sequelize);
       if (tableNames[0] && tableNames[0].tableName) {
         tableNames = tableNames.map(v => v.tableName);
       }
       expect(tableNames).to.deep.equal(['my_test_table']);
     });
 
-    if (dialect !== 'sqlite' && dialect !== 'postgres') {
+    if (!['sqlite', 'postgres', 'db2', 'oracle'].includes(dialect)) {
       // NOTE: sqlite doesn't allow querying between databases and
       // postgres requires creating a new connection to create a new table.
       it('should not show tables in other databases', async function() {
@@ -73,7 +82,7 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
       });
     }
 
-    if (dialect === 'mysql' || dialect === 'mariadb') {
+    if (['mysql', 'mariadb'].includes(dialect)) {
       it('should show all tables in all databases', async function() {
         await this.queryInterface.createTable('my_test_table1', { name: DataTypes.STRING });
         await this.sequelize.query('CREATE DATABASE my_test_db');
@@ -90,7 +99,9 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
           tableNames = tableNames.map(v => v.tableName);
         }
         tableNames.sort();
-        expect(tableNames).to.deep.equal(['my_test_table1', 'my_test_table2']);
+
+        expect(tableNames).to.include('my_test_table1');
+        expect(tableNames).to.include('my_test_table2');
       });
     }
   });
@@ -102,7 +113,7 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
       });
       await this.queryInterface.renameTable('my_test_table', 'my_test_table_new');
       let tableNames = await this.queryInterface.showAllTables();
-      if (dialect === 'mssql' || dialect === 'mariadb') {
+      if (['mssql', 'mariadb', 'db2', 'oracle'].includes(dialect)) {
         tableNames = tableNames.map(v => v.tableName);
       }
       expect(tableNames).to.contain('my_test_table_new');
@@ -144,7 +155,7 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
       });
       await this.queryInterface.dropAllTables({ skip: ['skipme'] });
       let tableNames = await this.queryInterface.showAllTables();
-      if (dialect === 'mssql' || dialect === 'mariadb') {
+      if (['mssql', 'mariadb', 'db2', 'oracle'].includes(dialect)) {
         tableNames = tableNames.map(v => v.tableName);
       }
       expect(tableNames).to.contain('skipme');
@@ -268,22 +279,24 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
       expect(table).to.not.have.property('active');
     });
 
-    it('renames a column primary key autoIncrement column', async function() {
-      const Fruits = this.sequelize.define('Fruit', {
-        fruitId: {
-          type: DataTypes.INTEGER,
-          allowNull: false,
-          primaryKey: true,
-          autoIncrement: true
-        }
-      }, { freezeTableName: true });
+    if (dialect !== 'db2') { // Db2 does not allow rename of a primary key column
+      it('renames a column primary key autoIncrement column', async function() {
+        const Fruits = this.sequelize.define('Fruit', {
+          fruitId: {
+            type: DataTypes.INTEGER,
+            allowNull: false,
+            primaryKey: true,
+            autoIncrement: true
+          }
+        }, { freezeTableName: true });
 
-      await Fruits.sync({ force: true });
-      await this.queryInterface.renameColumn('Fruit', 'fruitId', 'fruit_id');
-      const table = await this.queryInterface.describeTable('Fruit');
-      expect(table).to.have.property('fruit_id');
-      expect(table).to.not.have.property('fruitId');
-    });
+        await Fruits.sync({ force: true });
+        await this.queryInterface.renameColumn('Fruit', 'fruitId', 'fruit_id');
+        const table = await this.queryInterface.describeTable('Fruit');
+        expect(table).to.have.property('fruit_id');
+        expect(table).to.not.have.property('fruitId');
+      });
+    }
 
     it('shows a reasonable error message when column is missing', async function() {
       const Users = this.sequelize.define('_Users', {
@@ -369,18 +382,19 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
       });
       expect(table).to.have.property('level_id');
     });
-
-    it('should work with enums (1)', async function() {
-      await this.queryInterface.addColumn('users', 'someEnum', DataTypes.ENUM('value1', 'value2', 'value3'));
-    });
-
-    it('should work with enums (2)', async function() {
-      await this.queryInterface.addColumn('users', 'someOtherEnum', {
-        type: DataTypes.ENUM,
-        values: ['value1', 'value2', 'value3']
+    // Db2 does not support enums in alter column
+    if (dialect !== 'db2') {
+      it('should work with enums (1)', async function() {
+        await this.queryInterface.addColumn('users', 'someEnum', DataTypes.ENUM('value1', 'value2', 'value3'));
       });
-    });
 
+      it('should work with enums (2)', async function() {
+        await this.queryInterface.addColumn('users', 'someOtherEnum', {
+          type: DataTypes.ENUM,
+          values: ['value1', 'value2', 'value3']
+        });
+      });
+    }
     if (dialect === 'postgres') {
       it('should be able to add a column of type of array of enums', async function() {
         await this.queryInterface.addColumn('users', 'tags', {
@@ -446,7 +460,7 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
       const foreignKeys = await this.sequelize.query(
         this.queryInterface.queryGenerator.getForeignKeysQuery(
           'hosts',
-          this.sequelize.config.database
+          dialect === 'db2' ? this.sequelize.config.username.toUpperCase() : this.sequelize.config.database
         ),
         { type: this.sequelize.QueryTypes.FOREIGNKEYS }
       );
@@ -457,10 +471,12 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
         expect(Object.keys(foreignKeys[0])).to.have.length(6);
         expect(Object.keys(foreignKeys[1])).to.have.length(7);
         expect(Object.keys(foreignKeys[2])).to.have.length(7);
-      } else if (dialect === 'sqlite') {
+      } else if (dialect === 'sqlite' || dialect === 'db2') {
         expect(Object.keys(foreignKeys[0])).to.have.length(8);
-      } else if (dialect === 'mysql' || dialect === 'mariadb' || dialect === 'mssql') {
+      } else if (['mysql', 'mariadb', 'mssql'].includes(dialect)) {
         expect(Object.keys(foreignKeys[0])).to.have.length(12);
+      } else if (dialect === 'oracle') {
+        expect(Object.keys(foreignKeys[0])).to.have.length(6);
       } else {
         throw new Error(`This test doesn't support ${dialect}`);
       }
@@ -489,8 +505,10 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
   describe('constraints', () => {
     beforeEach(async function() {
       this.User = this.sequelize.define('users', {
-        username: DataTypes.STRING,
-        email: DataTypes.STRING,
+        // Db2 does not allow unique constraint for a nullable column, Db2
+        // throws SQL0542N error if we create constraint on nullable column.
+        username: dialect === 'db2' ? { type: DataTypes.STRING, allowNull: false } : DataTypes.STRING,
+        email: dialect === 'db2' ? { type: DataTypes.STRING, allowNull: false } : DataTypes.STRING,
         roles: DataTypes.STRING
       });
 
@@ -599,7 +617,7 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
         constraints = constraints.map(constraint => constraint.constraintName);
 
         // The name of primaryKey constraint is always `PRIMARY` in case of MySQL and MariaDB
-        const expectedConstraintName = dialect === 'mysql' || dialect === 'mariadb' ? 'PRIMARY' : 'users_username_pk';
+        const expectedConstraintName = ['mysql', 'mariadb'].includes(dialect) ? 'PRIMARY' : 'users_username_pk';
 
         expect(constraints).to.include(expectedConstraintName);
         await this.queryInterface.removeConstraint('users', expectedConstraintName);
@@ -627,7 +645,7 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
             field: 'username'
           },
           onDelete: 'cascade',
-          onUpdate: 'cascade',
+          onUpdate: dialect !== 'oracle' ? 'cascade' : null,
           type: 'foreign key'
         });
         let constraints = await this.queryInterface.showConstraint('posts');
@@ -649,8 +667,11 @@ describe(Support.getTestDialectTeaser('QueryInterface'), () => {
           throw new Error('Error not thrown...');
         } catch (error) {
           expect(error).to.be.instanceOf(Sequelize.UnknownConstraintError);
-          expect(error.table).to.equal('users');
-          expect(error.constraint).to.equal('unknown__constraint__name');
+          // The Oracle dialect, error messages doesn't have table and constraint information
+          if (dialect != 'oracle') {
+            expect(error.table).to.equal('users');
+            expect(error.constraint).to.equal('unknown__constraint__name');
+          }
         }
       });
     });
